bugfix on reentrant connection handler
There was a problem with connection_write due to handler
reentrancy (and consequently, borrow problems).

It happened because a write would immediatelly try to write.
Only on WouldBlock we would set the write interest.

The problem is that writing may generate calls to the
Handler (on_disconnect/on_write_finished), and thus, cause problems
when called from within one of the handler fns.

The solution was to register write interest instead of immediatelly
writing. The drawback is extra overhead: writes mostly don't return
WouldBlock, and we would avoid the register/poll when writing
immediately.